backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
aiohttp
reportlab
google-generativeai
PyGithub
python-multipart
chromadb
sentence-transformers
firebase-admin
aioredis
pydantic[email]
colorama
psutil
Pillow 